src/requirements.txt
kubernetes==26.1.0
boto3==1.28.29
credstash==1.17.1
prometheus-client==0.17.1
flask==2.1.0
waitress==2.0.0
flask-limiter==3.5.0
pytz==2023.3
cmake==3.27.2
google-cloud-bigtable==2.28.1
google-cloud-pubsub==2.28.0